fixes broken pNodes
Docker has moved from v1 to v2 registries. Because of this the start.sh script is broken. The start_incognito script is used as a template to create the start.sh file. The start.sh file is created by running main.py, but I have not tested doing this as I do not want to brick my node. A patch file can be created and run manually or you might have other means of pushing these changes out. I've also made a few commands in this script much simpler and more readable where possible.